pyproject.toml
[tool.poetry]
name = "sample"
version = "0.0.1"
description = ""
authors = ["MIERUNE Inc. <info@mierune.co.jp>"]
readme = "README.md"
packages = [
  #{ include = "qgis", from = "C:\\Program Files\\QGIS 3.28.2\\apps\\qgis\\python" },
]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.2.1"
flake8-bugbear = "^23.3.12"
pytest-cov = "^4.0.0"
pytest-qgis = "^1.3.2"


[tool.poetry.group.pyqt5.dependencies]
pyqt5 = "^5.15.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
